// old_dist/CommonTypes.d.ts
type RefType = string | ObjectId;
export type EventEmitter3 = EventEmitter;
export type AutoProps<T> = {
    readonly [K in keyof T]: T[K];
};
export type InstanceOf<T> = T extends Constructor<infer I> ? I : T;
export type Constructor<T> = new (...args: any[]) => T;
export type UnboxConstructor<T> = T extends new (...args: any[]) => infer I ? I : T;
export type LoggersType = {
    info: (s: string) => void;
    debug: (s: string) => void;
    error: (s: string) => void;
    warn: (s: string) => void;
};
type IsAUCO<T> = T extends {
    className: string;
} ? true : false;
type AllowStringForRefs<V> = NonNullable<V> extends Array<infer U> ? IsAUCO<NonNullable<U>> extends true ? (U | string | ObjectIdLike)[] : V : IsAUCO<NonNullable<V>> extends true ? V | string | ObjectIdLike : V;
type OnlyStringForRefs<V> = NonNullable<V> extends Array<infer U> ? IsAUCO<NonNullable<U>> extends true ? string[] : V : IsAUCO<NonNullable<V>> extends true ? string : V;
export type Cache<T> = {
    references: {
        [K in keyof T]?: AutoUpdateManager<AutoUpdatedClientObject<any>>;
    };
};
export type IsData<T> = {
    [K in keyof T]: AllowStringForRefs<T[K]>;
} & {
    _id: any;
};
export type ExtractedData<T, Base = AutoUpdatedClientObject<T>> = {
    [K in keyof FixPure<T, Base>]: OnlyStringForRefs<T[K]>;
};
type FixPure<T, Base> = Omit<T, keyof Omit<Base, "_id">>;
export type SocketEvent = [string, any, (res: ServerResponse<any>) => void];
export type ServerResponse<T> = {
    data: T;
    message?: string;
    success: true;
} | {
    message: string;
    success: false;
};
export type ServerUpdateRequest<T> = {
    _id: RefType;
    key: string;
    value: any;
};
// ...
export type OnlyAddedKeys<Sub, Parent> = Pick<Sub, Exclude<keyof Sub, keyof Omit<Parent, "_id">>>;
export type Prev = [never, 0, 1, 2, 3];
export type Join<K, P> = K extends string | number ? P extends string | number ? `${K}${"" extends P ? "" : "."}${P}` : never : never;
export type Paths<T, Base, D extends number = 3> = [D] extends [never] ? never : NonNullable<T> extends object ? {
    [K in keyof OnlyAddedKeys<NonNullable<T>, Base> & string]-?: NonNullable<NonNullable<T>[K]> extends Function ? never : NonNullable<NonNullable<T>[K]> extends Array<any> | Date | ObjectId ? K : NonNullable<NonNullable<T>[K]> extends object ? K | Join<K, Paths<NonNullable<NonNullable<T>[K]>, Base, Prev[D]>> : K;
}[keyof OnlyAddedKeys<NonNullable<T>, Base> & string] : never;
export type PathValueOf<T, P extends string> = P extends `${infer K}.${infer Rest}` ? K extends keyof NonNullable<T> ? PathValueOf<NonNullable<NonNullable<T>[K]>, Rest> : never : P extends keyof NonNullable<T> ? NonNullable<T>[P] : never;